requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dlib==19.24.6
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11.1
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
